Resolves DCP2-482: ArcLight - Create Short URL for each Repository with good content
